requirements.txt
aiofiles
aiohttp
apscheduler
cairosvg
cloudscraper
colour
covid
cowpy
dataclasses
DateTime
emoji==1.7.0
fonttools
geopy
gitpython
glitch_this
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gtts
hachoir
heroku3
html-telegraph-poster
httpx[http2]
humanize
IMDbPY
jikanpy
justwatch
lottie
lyricsgenius
markdown
markdown-it-py
motor
moviepy
nekos.py
opencv-python-headless
Pillow
prettytable
psutil
psycopg2
pyfiglet
PyGithub
pygments
pylast
pymediainfo
PyMuPDF
pySmartDL
python-barcode
python-dotenv
pytz
qrcode
randomstuff.py
regex
requests
search-engine-parser
selenium
setuptools
ShazamAPI
spamwatch
speedtest-cli
sqlalchemy-json
sqlalchemy==1.3.23
telegraph
tgcrypto
ujson
urlextract
validators
vcsi
wand
wget
yt_dlp
youtube_search
fake_useragent
faker
htmlwebshot
Telethon==1.25.0

git+https://github.com/alexmercerind/youtube-search-python
git+https://github.com/ITS-LEGENDBOT/py-googletrans
git+https://github.com/goldsmith/Wikipedia

